Rename file to get rid of CodeMirror vocabulary too
